Ref T21270: Eliminated compiler warnings where possible.
Summary:
Fixed warnings in graph_output_0_ip_test.cpp

Fixed warnings in numerics_ip_nip_0_test.cpp

Fixed warnings in numerics_ip_nip_1_test.cpp

Fixed warnings in stepio_nelms_error_test.cpp

Fixed warnings in graphtransformer_impl.cpp

Fixed warning in expandx.cpp

Fixed warnings in identityx.cpp

Fixed warnings in identityx.cpp (previous fix didn't work)

Fixed warning in ir.cpp (making explicit which switch cases we're missing in Ir::growLossGradients).

Warning fixes in cachesetup.cpp

Warning fixes in elementwise.cpp

Warning fixes in nll.cpp

Fixed warnings in aliaszerocopy.cpp

Fixed warnings in softmax.cpp

Fixed warning in devicex.cpp.

Fixed warning in subgraphutil.hpp.

Fixed warnings in maxclique.hpp

Fixed warnings in maxclique.cpp

Fixed warnings in onnxconstexpr.cpp

Fixed warnings in sgd0decompose.cpp

Fixed warnings in nllx.cpp

Fixed warnings in onnxconstexpr.cpp

Fixed warnings for identityx.cpp

Fixed warnings in l1x.cpp

Fixed warnings in identityx.cpp

Fixed warning in tensornames.cpp

Fixing warning in aliases.hpp

Fixed warning in cachesetup.cpp.

Fixed warning in batchserialize.cpp

Fixed warning in floorce.cpp

Fixed warning in batchserialise.cpp

Added static assertion to graphtransformer_iml.cpp

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, anthonyb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn

Subscribers: matthewha

Maniphest Tasks: T21270, T20598

Differential Revision: https://phabricator.sourcevertex.net/D28156

+++ willow/src/popx/op/expandx.cpp
@@ -33,7 +33,7 @@ poplar::Tensor BaseExpandOpx::unwindTensorLayout(poplar::Tensor tensor,
         tensor =
             tensor.slice(static_cast<std::size_t>(0),
                          static_cast<std::size_t>(input_shape_of_current_dim),
-                         static_cast<std::size_t>(dim));
+                         static_cast<unsigned int>(dim));
       }
     }
   }
